create tests for App container
